Move dbTypes to shared folder
Frontend now also has access to shared types. API types will also be
put here.

Frontend compiler doesn't seem to be able to successfully watch the
dbTypes file. Maybe only one compiler can watch a file at a time? The
backend compiler is able to watch dbTypes. Is it possible these two
will switch? Possibly depending on which was started later/earlier?

For now, when editing shared files, make sure to manually trigger
recompilation in both compilers.
