+++ client/src/Playlist-Editor/Chat/SeparateChat.tsx
@@ -3,14 +3,8 @@ import React, { CSSProperties } from 'react'
 import { classes } from '../../styles'
 // import * as styles from './playlistTableRowStyles'
 // import { State } from './modificationReducer'
-import { SeparateMessageEditor } from '../MessageEditor'
+import { SeparateMessageEditor } from './MessageEditor'
 
-
-
-// const chatStyle = {
-//   ...classes.column,
-//   // margin: '0 1.0rem 1.0rem',
-// }
 
 
 export const SeparateChat = ({
